feat: service code 구현

+++ src/main/java/umc/spring_study/web/dto/MemberDTO/MemberResponseDTO.java
@@ -27,7 +27,7 @@ public class MemberResponseDTO {
         String foodCategoryName;
         String missionSpec;
         Integer reward;
-        Integer deadline;
+        Long deadline;
     }
 
     @Schema(description="사용자 홈화면 - 참여가능한 미션")
